pyproject.toml
[project]
name = "codingbuddy"
version = "0.1.0"
description = "AI-powered side project generator using Claude agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "typer>=0.12.0",
    "httpx>=0.27.0",
    "PyGithub>=2.4.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "websockets>=12.0",
    "aiofiles>=23.0.0",
    "rich>=13.0.0",
    "gitpython>=3.1.0",
]

[project.scripts]
buddy = "buddy.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["buddy", "dashboard"]
